horodocs_front/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-health-check==3.17.0
gunicorn==20.1.0
idna==3.4
psycopg2==2.9.6
python-dotenv==1.0.0
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
